# backend/src/api/bridge-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Robot Control System - Device Bridge",
    "description": "Python bridge server for device communication",
    "version": "1.0.0"
  },
  "paths": {
...
    "/devices/{device_id}/diagnostics": {
      "post": {
        "summary": "Run Diagnostics",
        "description": "Board diagnosztika futtatása a meglévő serial kapcsolaton.",
        "operationId": "run_diagnostics_devices__device_id__diagnostics_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "move_test",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Move Test"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/firmware-probe": {
      "post": {
        "summary": "Run Firmware Probe",
        "description": "Firmware paraméterek felderítése – különböző parancsok kipróbálása.",
        "operationId": "run_firmware_probe_devices__device_id__firmware_probe_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/endstop-test": {
      "post": {
        "summary": "Run Endstop Test",
        "description": "Végállás teszt – minden tengely végállásig mozgatása.",
        "operationId": "run_endstop_test_devices__device_id__endstop_test_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "step_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 5.0,
              "title": "Step Size"
            }
          },
          {
            "name": "speed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 15,
              "title": "Speed"
            }
          },
          {
            "name": "max_angle",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 200.0,
              "title": "Max Angle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/endstops": {
      "get": {
        "summary": "Get Endstop States",
        "description": "Végállás érzékelők aktuális állapotának lekérdezése (M119).",
        "operationId": "get_endstop_states_devices__device_id__endstops_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/motion-test": {
      "post": {
        "summary": "Run Motion Test",
        "description": "Mozgásminőség teszt – különböző sebességekkel.",
        "operationId": "run_motion_test_devices__device_id__motion_test_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "test_angle",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 30.0,
              "title": "Test Angle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/cancel-test": {
      "post": {
        "summary": "Cancel Test",
        "description": "Futó teszt (firmware-probe, endstop-test, motion-test) leállítása.",
        "operationId": "cancel_test_devices__device_id__cancel_test_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/test-progress": {
      "get": {
        "summary": "Get Test Progress",
        "description": "Futó teszt napló lekérdezése (polling). after = ennyi bejegyzést ugorjon át.",
        "operationId": "get_test_progress_devices__device_id__test_progress_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}